Add files via upload
This is a simple test script used to see if the memory scanner program works. Created a variable of integer value which is printed every few seconds. The end goal is to run both programs and change the integer value of the variable to see the memory being modified in real time.